+++ packages/tools/src/tools/annotation/RectangleROITool.ts
@@ -24,7 +24,7 @@ import {
 import {
   drawHandles as drawHandlesSvg,
   drawLinkedTextBox as drawLinkedTextBoxSvg,
-  drawRect as drawRectSvg,
+  drawRectByCoordinates as drawRectSvg,
 } from '../../drawingSvg';
 import { state } from '../../store';
 import { Events } from '../../enums';
@@ -171,6 +171,7 @@ class RectangleROITool extends AnnotationTool {
         viewUp: <Types.Point3>[...viewUp],
         FrameOfReferenceUID,
         referencedImageId,
+        ...viewport.getViewReference({ points: [worldPos] }),
       },
       data: {
         label: '',
@@ -748,8 +749,7 @@ class RectangleROITool extends AnnotationTool {
         svgDrawingHelper,
         annotationUID,
         rectangleUID,
-        canvasCoordinates[0],
-        canvasCoordinates[3],
+        canvasCoordinates,
         {
           color,
           lineDash,