+++ packages/tools/src/tools/displayTools/Contour/contourDisplay.ts
@@ -1,7 +1,7 @@
 import {
   getEnabledElementByIds,
+  StackViewport,
   Types,
-  BaseVolumeViewport,
 } from '@cornerstonejs/core';
 
 import Representations from '../../../enums/SegmentationRepresentations';
@@ -11,10 +11,10 @@ import {
   SegmentationRepresentationConfig,
   ToolGroupSpecificRepresentation,
 } from '../../../types/SegmentationStateTypes';
-import { addOrUpdateVTKContourSets } from './vtkContour/addOrUpdateVTKContourSets';
 import removeContourFromElement from './removeContourFromElement';
-import { deleteConfigCache } from './vtkContour/contourConfigCache';
+import { deleteConfigCache } from './contourHandler/contourConfigCache';
 import { polySeg } from '../../../stateManagement/segmentation';
+import { handleContourSegmentation } from './contourHandler/handleContourSegmentation';
 
 let polySegConversionInProgress = false;
 
@@ -63,7 +63,7 @@ function removeSegmentationRepresentation(
  * @param toolGroupConfig - This is the configuration object for the tool group
  */
 async function render(
-  viewport: Types.IVolumeViewport,
+  viewport: StackViewport | Types.IVolumeViewport,
   representationConfig: ToolGroupSpecificRepresentation,
   toolGroupConfig: SegmentationRepresentationConfig
 ): Promise<void> {
@@ -95,89 +95,19 @@ async function render(
     );
   }
 
-  // From here to below it is basically the legacy geometryId based
-  // contour rendering via vtkActors that has some bugs for display,
-  // as it sometimes appear and sometimes not, and it is not clear.
-  // We have moved to the new SVG based contours via our annotation tools
-  // check out annotationUIDsMap in the ContourSegmentationData type
-  const { geometryIds } = contourData;
-
-  if (!geometryIds?.length || !(viewport instanceof BaseVolumeViewport)) {
+  if (!contourData) {
     return;
   }
 
-  // add the contour sets to the viewport
-  addOrUpdateVTKContourSets(
-    viewport,
-    geometryIds,
-    representationConfig,
-    toolGroupConfig
-  );
-
-  /**
-   * The following logic could be added if we want to support the use case
-   * where the contour representation data is initiated using annotations
-   * in the state from the get-go , and not when the user draws a contour.
-   */
-  // if (contourData?.points?.length) {
-  //   // contourData = createAnnotationsFromPoints(contourData.points);
-  //   const contourSegmentationAnnotation = {
-  //     annotationUID: csUtils.uuidv4(),
-  //     data: {
-  //       contour: {
-  //         closed: true,
-  //         polyline: contourData.points,
-  //       },
-  //       segmentation: {
-  //         segmentationId,
-  //         segmentIndex: 1, // Todo
-  //         segmentationRepresentationUID:
-  //           representationConfig.segmentationRepresentationUID,
-  //       },
-  //     },
-  //     highlighted: false,
-  //     invalidated: false,
-  //     isLocked: false,
-  //     isVisible: true,
-  //     metadata: {
-  //       toolName: 'PlanarFreehandContourSegmentationTool',
-  //       FrameOfReferenceUID: viewport.getFrameOfReferenceUID(),
-  //       viewPlaneNormal: viewport.getCamera().viewPlaneNormal,
-  //     },
-  //   };
-
-  //   addAnnotation(contourSegmentationAnnotation, viewport.element);
-  // } else if (
-  //   !contourData &&
-  //   polySeg.canComputeRequestedRepresentation(
-  //     representationConfig.segmentationRepresentationUID
-  //   )
-  // ) {
-  // contourData = await polySeg.computeAndAddContourRepresentation(
-  //   segmentationId,
-  //   {
-  //     segmentationRepresentationUID:
-  //       representationConfig.segmentationRepresentationUID,
-  //     viewport,
-  //   }
-  // );
-  // }
-
-  // if (contourData?.geometryIds?.length) {
-  //   handleVTKContour({
-  //     viewport,
-  //     representationConfig,
-  //     toolGroupConfig,
-  //     geometryIds: contourData.geometryIds,
-  //   });
-  // } else if (contourData.annotationUIDsMap?.size) {
-  //   handleContourAnnotationSegmentation({
-  //     viewport,
-  //     representationConfig,
-  //     toolGroupConfig,
-  //     annotationUIDsMap: contourData.annotationUIDsMap,
-  //   });
-  // }
+  if (contourData?.geometryIds?.length) {
+    handleContourSegmentation(
+      viewport,
+      contourData.geometryIds,
+      contourData.annotationUIDsMap,
+      representationConfig,
+      toolGroupConfig
+    );
+  }
 }
 
 function _removeContourFromToolGroupViewports(
@@ -198,10 +128,7 @@ function _removeContourFromToolGroupViewports(
       viewportId,
       renderingEngineId
     );
-    removeContourFromElement(
-      enabledElement.viewport.element,
-      segmentationRepresentationUID
-    );
+    removeContourFromElement(segmentationRepresentationUID, toolGroupId);
   }
 }