+++ packages/tools/src/utilities/index.ts
@@ -28,7 +28,10 @@ import { pointToString } from './pointToString';
 import annotationFrameRange from './annotationFrameRange';
 import pointInSurroundingSphereCallback from './pointInSurroundingSphereCallback';
 import getViewportForAnnotation from './getViewportForAnnotation';
-
+import {
+  annotationHydration,
+  getClosestImageIdForStackViewport,
+} from './annotationHydration';
 // name spaces
 import * as contours from './contours';
 import * as segmentation from './segmentation';
@@ -97,4 +100,6 @@ export {
   voi,
   annotationFrameRange,
   contourSegmentation,
+  annotationHydration,
+  getClosestImageIdForStackViewport,
 };