+++ tests/utils/screenShotPaths.ts
@@ -1,0 +1,50 @@
+/**
+ * Paths to the screenshots of the tests.
+ */
+const screenShotPaths = {
+  stackBasic: {
+    viewport: 'viewport.png',
+  },
+  stackAPI: {
+    setVoiRange: 'setVoiRange.png',
+    nextImage: 'nextImage.png',
+    previousImage: 'previousImage.png',
+    flipH: 'flipH.png',
+    flipV: 'flipV.png',
+    // rotateRandom: 'rotateRandom.png',
+    rotateAbsolute150: 'rotateAbsolute150.png',
+    rotateDelta30: 'rotateDelta30.png',
+    invert: 'invert.png',
+    // randomZoomAndPan: 'randomZoomAndPan.png',
+    colormap: 'colormap.png',
+    resetViewport: 'resetViewport.png',
+  },
+  stackManipulationTools: {
+    planarRotate: 'planarRotate.png',
+    windowLevel: 'windowLevel.png',
+  },
+  stackProperties: {
+    nextImage: 'nextImage.png',
+    previousImage: 'previousImage.png',
+    propertiesAddedForCurrentImage: 'propertiesAddedForCurrentImage.png',
+    propertiesAreSameForNextImage: 'propertiesAreSameForNextImage.png',
+    propertiesRemovedForCurrentImage: 'propertiesRemovedForCurrentImage.png',
+    propertiesAreSameForPreviousImage: 'propertiesAreSameForPreviousImage.png',
+    resetToDefaultViewportProperties: 'resetToDefaultViewportProperties.png',
+    resetMetadata: 'resetMetadata.png',
+  },
+  volumeBasic: {
+    viewport: 'viewport.png',
+  },
+  ultrasoundColors: {
+    slice1: 'slice1.png',
+    slice2: 'slice2.png',
+    slice3: 'slice3.png',
+    slice4: 'slice4.png',
+    slice5: 'slice5.png',
+    slice6: 'slice6.png',
+    slice7: 'slice7.png',
+  },
+};
+
+export { screenShotPaths };